# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin-dashboard/statistics": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get system statistics",
                "operationId": "63209874b9061fab1f035dfff09b5327",
                "responses": {
                    "200": {
                        "description": "Statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_borrows": {
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "total_returns": {
                                            "type": "integer",
                                            "example": 120
                                        },
                                        "overdue_borrows": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "current_total_books": {
                                            "type": "integer",
                                            "example": 300
                                        },
                                        "new_members": {
                                            "type": "integer",
                                            "example": 25
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-dashboard/borrows-chart": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get number of borrows per day for the last 7 days",
                "operationId": "781c02277d07c2ef88f4b8a5bdee196e",
                "responses": {
                    "200": {
                        "description": "Successful response with daily borrow counts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "day": {
                                                "type": "string",
                                                "example": "Monday"
                                            },
                                            "total": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-dashboard/recent-borrows": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get recent borrows in the last 7 days (paginated)",
                "description": "Returns a paginated list of recent borrows within the past week.",
                "operationId": "getRecentBorrows",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of records per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful paginated response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Borrow"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://yourdomain.com/api/borrows/recent?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://yourdomain.com/api/borrows/recent?page=5"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://yourdomain.com/api/borrows/recent?page=2",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://yourdomain.com/api/borrows/recent"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-dashboard/overdue-borrows": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get list of overdue borrowed books",
                "operationId": "44ac62145bb3a8589376f18fca1c9949",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Overdue borrowed books list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Borrow"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-dashboard/new-borrows": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get list of new borrow requests",
                "operationId": "e2c41d3884da185f0894c7faef3734c7",
                "responses": {
                    "200": {
                        "description": "New borrow requests list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Borrow"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-dashboard/borrows/{id}/approve": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Approve a borrow request",
                "operationId": "d544a870f6d46f1d499f510cc71cb010",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Borrow request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrow approved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Borrow approved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Borrow request is not pending",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Borrow request is not pending"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Borrow request not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Borrow request not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Borrow": {
                "required": [
                    "id",
                    "borrowed_at",
                    "return_date",
                    "status",
                    "extension_count"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "borrowed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-01T10:00:00Z"
                    },
                    "return_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-31T23:59:59Z"
                    },
                    "status": {
                        "type": "string",
                        "example": "borrowed"
                    },
                    "returned_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-25T14:30:00Z",
                        "nullable": true
                    },
                    "extension_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "book": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Book"
                            }
                        ],
                        "nullable": true
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "role"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "customer"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-15"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St, Dhaka, Bangladesh"
                    },
                    "profile_picture": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/users/john.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}